moved all test to _test_ folder and edited test fiels
