+++ frontend/src/_test_/AddUser.test.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import UsersTableHead from '../../Components/Users/UsersTableHead';
+import AddUser from '../Components/Users/AddUser';
 
 configure({ adapter: new Adapter() });
 
-describe('UsersTableHead Component', () => {
+describe('AddUser Component', () => {
     it('Should render without crashing', () => {
         const wrapper = 
-        shallow(<UsersTableHead />).length
+        shallow(<AddUser />).length
         expect(wrapper).toEqual(1)
     })
 })
