+++ frontend/src/_test_/Notification.test.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import AddUser from '../../Components/Users/AddUser';
+import Notification from '../Components/Users/Notification';
 
 configure({ adapter: new Adapter() });
 
-describe('AddUser Component', () => {
+describe('Notification Component', () => {
     it('Should render without crashing', () => {
         const wrapper = 
-        shallow(<AddUser />).length
+        shallow(<Notification />).length
         expect(wrapper).toEqual(1)
     })
 })
