+++ frontend/src/_test_/UserDropDown.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import UserDropDown from '../../Components/Users/UserDropDown';
+import UserDropDown from '../Components/Users/UserDropDown';
 
 configure({ adapter: new Adapter() });
