+++ frontend/src/_test_/UsersTableHead.test.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Notification from '../../Components/Users/Notification';
+import UsersTableHead from '../Components/Users/UsersTableHead';
 
 configure({ adapter: new Adapter() });
 
-describe('Notification Component', () => {
+describe('UsersTableHead Component', () => {
     it('Should render without crashing', () => {
         const wrapper = 
-        shallow(<Notification />).length
+        shallow(<UsersTableHead />).length
         expect(wrapper).toEqual(1)
     })
 })
